requirements.txt
fitz==0.0.1.dev2
gTTS==2.5.3
protobuf>=3.20,<5.0.0
python-dotenv==1.0.1
SpeechRecognition==3.10.4
google-generativeai==0.7.2
frontend==0.0.3
PyPdf2==3.0.1